requirements.txt
amqp==2.3.2
aniso8601==4.0.1
asn1crypto==0.24.0
Babel==2.6.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
bs4==0.0.1
celery==4.2.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Django==1.8.2
dnspython==1.15.0
flasgger==0.9.2
Flask==1.0.2
Flask-Docs==0.0.8
Flask-RESTful==0.3.6
flower==0.9.2
future==0.16.0
grpcio==1.15.0
grpcio-tools==1.15.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.1
MarkupSafe==1.1.0
meld3==1.0.2
mistune==0.8.4
protobuf==3.6.1
pycparser==2.18
PyMySQL==0.9.2
pysnowflake==0.1.3
python-consul==1.1.0
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
six==1.11.0
supervisor==4.0.0.dev0
tornado==5.1.1
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4
Werkzeug==0.14.1
